Split up InMemoryAuthority
